requirements.txt
Django==3.2.12
lzstring==1.0.4
pycryptodome==3.18.0